pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "AI projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-storage-blob>=12.28.0",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.0",
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "openai>=1.91.0",
    "opencensus-ext-azure>=1.1.15",
    "pydantic>=2.11.7",
    "pypdf>=5.6.1",
    "reportlab>=4.4.4",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "projects",
]

[tool.uv.sources]
projects = { workspace = true }
